requirements.txt
boto3==1.26.59
botocore==1.29.59
jmespath==1.0.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
requests==2.28.2
jaydebeapi==1.2.3
SQLAlchemy==2.0.9
pandas==2.0.0
DateTime==5.1
fastapi==0.95.1
uvicorn==0.21.1